+++ src/lib/audit.js
@@ -415,7 +415,16 @@ const auditTenant = function (context, tenantId, declaration, commonConfig, prev
         .then(() => fetch.checkDesiredForReferencedProfiles(context, tenantDesiredConfig))
         .then(() => {
             if (tenantId === 'Common') {
-                return fetch.getCommonAccessProfiles(context);
+                const dataGroupInfo = {
+                    name: 'commonAccessProfiles',
+                    storageName: 'accessProfiles'
+                };
+                return fetch.getDataGroupData(context, dataGroupInfo)
+                    .then(() => {
+                        dataGroupInfo.name = 'commonRouteDomains';
+                        dataGroupInfo.storageName = 'routeDomains';
+                        return fetch.getDataGroupData(context, dataGroupInfo);
+                    });
             }
             return Promise.resolve();
         })
@@ -489,7 +498,18 @@ const auditTenant = function (context, tenantId, declaration, commonConfig, prev
         })
         .then((response) => {
             if (tenantId === 'Common' && !context.tasks[context.currentIndex].firstPassNoDelete) {
-                return fetch.updateCommonAccessProfiles(context, tenantDesiredConfig)
+                const dataGroupInfo = {
+                    command: 'apm profile access',
+                    name: 'commonAccessProfiles',
+                    storageName: 'accessProfiles'
+                };
+                return fetch.updateDataGroup(context, tenantDesiredConfig, dataGroupInfo)
+                    .then(() => {
+                        dataGroupInfo.name = 'commonRouteDomains';
+                        dataGroupInfo.storageName = 'routeDomains';
+                        dataGroupInfo.command = 'net route-domain';
+                        return fetch.updateDataGroup(context, tenantDesiredConfig, dataGroupInfo);
+                    })
                     .then(() => response);
             }
             return response;