+++ test/unit/lib/map_cli.js
@@ -2142,4 +2142,57 @@ describe('map_cli', () => {
             assert.deepStrictEqual(apmUpdates, expectedUpdates);
         });
     });
+
+    describe('net route-domain', () => {
+        it('should get modify command for route-domain', () => {
+            const diff = {
+                kind: 'N',
+                path: ['/Common/10', 'properties', 'fw-enforced-policy'],
+                rhs: '/Common/shared/firewallPolicy',
+                command: 'net route-domain',
+                lhsCommand: 'net route-domain',
+                rhsCommand: 'net route-domain'
+            };
+            const targetConfig = {
+                'fw-enforced-policy': '/Common/Shared/firewallPolicy'
+            };
+
+            const results = mapCli.tmshCreate(context, diff, targetConfig);
+            assert.deepStrictEqual(
+                results,
+                {
+                    preTrans: [],
+                    commands: [
+                        'tmsh::modify net route-domain /Common/10 fw-enforced-policy /Common/Shared/firewallPolicy'
+                    ],
+                    postTrans: [],
+                    rollback: []
+                }
+            );
+        });
+
+        it('should revert with a modify when diff kind is D', () => {
+            const diff = {
+                kind: 'D',
+                path: ['/Common/10', 'properties', 'fw-enforced-policy'],
+                rhs: '/Common/shared/firewallPolicy',
+                command: 'net route-domain',
+                lhsCommand: 'net route-domain',
+                rhsCommand: 'net route-domain'
+            };
+
+            const results = mapCli.tmshDelete(context, diff);
+            assert.deepStrictEqual(
+                results,
+                {
+                    preTrans: [],
+                    commands: [
+                        'tmsh::modify net route-domain /Common/10 fw-enforced-policy none'
+                    ],
+                    postTrans: [],
+                    rollback: []
+                }
+            );
+        });
+    });
 });