+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changes to this project are documented in this file. More detail (including info
 - AUTOTOOL-4094: Add the declaration id to the restnoded log output upon success and include it in the rest response
 - AUTOTOOL-4092: Expose failureRcode properties for GSLB WideIP
 - AUTOTOOL-4117: ([GitHub Issue 779](https://github.com/F5Networks/f5-appsvcs-extension/issues/779)): GSLB_Monitor HTTP/HTTPS receiveStatusCodes property
+- AUTOTOOL-4111: ([GitHub Issue 321](https://github.com/F5Networks/f5-declarative-onboarding/issues/321)): Add routeDomainEnforcement to Firewall_Policy to allow enforcing the Firewall_Policy on the specified Route Domain's
 
 ### Fixed
 - AUTOTOOL-4017: All pool member statuses reset to unknown when new member is added to the pool